requirements.txt
requests~=2.31.0
crypto~=1.4.1
pycryptodome~=3.21.0
cryptography~=41.0.7
dnspython~=2.3.0